Attribute Macro openbrush_lang_macro::trait_definition

#[trait_definition]

Defines extensible trait in the scope of openbrush::contract. It is a common rust trait, so you can use any features of rust inside of this trait. If this trait contains some methods marked with #[ink(message)] or #[ink(constructor)] attributes, this macro will extract these attributes and will put them into a separate trait (the separate trait only is used to call methods from the original trait), but the macro will not touch methods.

This macro stores definition of the trait in a temporary file during build process. Based on this definition #[openbrush::contract] will generate implementation of additional traits.

** Note ** The name of the trait defined via this macro must be unique for the whole project. ** Note ** You can’t use aliases, generics, and other rust’s stuff in signatures of ink!’s methods.

Example: Definition

mod doc {
use ink::prelude::collections::BTreeMap;
use openbrush::traits::{ AccountId, Balance, Storage, OccupyStorage };

#[derive(Debug)]
pub struct Data {
    pub balances: BTreeMap<AccountId, Balance>,
}

impl OccupyStorage for Data {
    const KEY: u32 = 0x123;
}

#[openbrush::trait_definition]
pub trait PSP22: Storage<Data> {
    /// Returns the account Balance for the specified `owner`.
    #[ink(message)]
    fn balance_of(&self, owner: AccountId) -> Balance {
        self.data().balances.get(&owner).copied().unwrap_or(0)
    }

    /// Transfers `value` amount of tokens from the caller's account to account `to`.
    #[ink(message)]
    fn transfer(&mut self, to: AccountId, value: Balance) {
        self._transfer_from_to(to, to, value)
    }

    fn _transfer_from_to(&mut self, from: AccountId, to: AccountId, amount: Balance) {
        let from_balance = self.balance_of(from);
        assert!(from_balance >= amount, "InsufficientBalance");
        self.data().balances.insert(from, from_balance - amount);
        let to_balance = self.balance_of(to);
        self.data().balances.insert(to, to_balance + amount);
    }
}
}

Example: Implementation

#[openbrush::contract]
mod base_psp22 {
    use ink::prelude::collections::BTreeMap;
    use openbrush::traits::Storage;

    const STORAGE_KEY: u32 = 123;

    #[derive(Default, Debug)]
    #[openbrush::upgradeable_storage(STORAGE_KEY)]
    pub struct Data {
        pub supply: Balance,
        pub balances: BTreeMap<AccountId, Balance>,
        pub allowances: BTreeMap<(AccountId, AccountId), Balance>,
    }

    #[openbrush::trait_definition]
    pub trait PSP22Example: Storage<Data> {
        /// Returns the account Balance for the specified `owner`.
        #[ink(message)]
        fn balance_of(&self, owner: AccountId) -> Balance {
            self.data().balances.get(&owner).copied().unwrap_or(0)
        }

        /// Transfers `value` amount of tokens from the caller's account to account `to`.
        #[ink(message)]
        fn transfer(&mut self, to: AccountId, value: Balance) {
            let from = Self::env().caller();
            self._transfer_from_to(from, to, value)
        }

        fn _transfer_from_to(&mut self, from: AccountId, to: AccountId, amount: Balance) {
            let from_balance = self.balance_of(from);
            assert!(from_balance >= amount, "InsufficientBalance");
            self.data().balances.insert(from, from_balance - amount);
            let to_balance = self.balance_of(to);
            self.data().balances.insert(to, to_balance + amount);
        }
    }

    #[ink(storage)]
    #[derive(Storage)]
    pub struct PSP22Struct {
        #[storage_field]
        example: Data,
        hated_account: AccountId,
    }

    impl Default for PSP22Struct {
        fn default() -> Self {
            Self {
                example: Data::default(),
                hated_account: AccountId::from([0x0; 32]),
            }
        }
    }

    impl PSP22Example for PSP22Struct {
        // Let's override method to reject transactions to bad account
        fn _transfer_from_to(&mut self, from: AccountId, to: AccountId, amount: Balance) {
            assert!(to != self.hated_account, "I hate this account!");

            let from_balance = self.balance_of(from);
            assert!(from_balance >= amount, "InsufficientBalance");
            self.get_mut().balances.insert(from, from_balance - amount);
            let to_balance = self.balance_of(to);
            self.get_mut().balances.insert(to, to_balance + amount);
        }
    }

    impl PSP22Struct {
        #[ink(constructor)]
        pub fn new(hated_account: AccountId) -> Self {
            let mut instance = Self::default();
            instance.hated_account = hated_account;
            instance
        }
    }
}